fix disk metrics calculation and reorganize code
